+++ src/common/js/startup.js
@@ -3,7 +3,7 @@
  */
 
 Meteor.startup(() => {
-    if( pwixI18n.conf.verbosity & PI_VERBOSE_DUMP ){
+    if( pwixI18n._conf.verbosity & PI_VERBOSE_DUMP ){
         console.debug( 'pwixI18n', pwixI18n );
     }
 });