+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - Replace merge dependency with lodash
     - configure() now acts both as a getter and a setter
+    - Rename conf to _conf, making clearer that this is a private data
 
 ### 1.3.2